requirements.txt
humanize>=0.5.1
slacker>=0.9.30
python-digitalocean>=1.10.1
